editviewoverlay: corrected AttributeChange
EditView on Overlay visualization needs to handle
selection change when e.g. FontSize attribute of
a selected text portion changes. Changed handling
so that selection overlay is an integral part of
text visualization overlay and refreshing text
visualization always refreshes selection visuals,
but not the other way around. Both anyways check
for real change, so just extra testing is avoided
which is still the better solution

Change-Id: I3ce6570fd9a18a7c1d7ced4bc4ca9aeda57538e0
Reviewed-on: https://gerrit.libreoffice.org/40993
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Armin Le Grand <Armin.Le.Grand@cib.de>