tdf#85909: EDITING: Native tables visualization not refreshed
... after changes to borders

Regression from:
4ea9c915ad7dd93f230d1775ecb6d25edda22408
"avoid repeated table layouting (fdo#75622)"

It was a performance change, but it was a bad idea to avoid
table layout refresh on this way

Anyway I added a call for border update even if not all
the table layouting is done. I tested with the test
document attached to fdo#75622, import time seems similar
so it does not cause perfromance issue to do that.

Change-Id: I7c6fcf105c89233512390dc2ecbd111a32f6779a
Reviewed-on: https://gerrit.libreoffice.org/41116
Reviewed-by: Tamás Zolnai <tamas.zolnai@collabora.com>
Tested-by: Tamás Zolnai <tamas.zolnai@collabora.com>