tdf#109142: Update to HarfBuzz 1.4.8
Should fix the Tibetan long U issue.

Also reverts this commit as depending on auto-detection for a required
feature is not a very good idea. The underlying issue is fixed upstream:

commit dad4e5ab9f4a5843ea05394abf0ea24c0510bb51
Author: jan Iversen <jani@libreoffice.org>
Date:   Mon Jun 12 14:51:42 2017 +0200

    iOS, harfbuzz libcoretext not found

    changed --with-libcoretext
    from "yes" to "auto" for IOS and MACOSX

which seems to be a misguided attemp to fix the build breakage
introduced in:

commit 6e79697608563f1f34e16c8c6aac3aff22178a5d
Author: jan Iversen <jani@libreoffice.org>
Date:   Sun Jun 11 19:35:53 2017 +0200

    iOS, patch harfbuzz

    Add support for arm64

Also had to drop ubsan.patch as it no longer applies, hopefully it is no
longer needed, otherwise it is always good idea to submit such patches
upstream.

Change-Id: I91643c542c0a5a00f36d5564052787eedef8c9cf
Reviewed-on: https://gerrit.libreoffice.org/41124
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Michael Stahl <mstahl@redhat.com>