tdf#99692 sw: fix the fix: don't set bNullContent if node was split
When showing the redlines in rhbz908615-13.odt, the following assertion
happens:

Assertion `IDocumentMarkAccess::IsLegalPaMForCrossRefHeadingBookmark(rPaM) && "<CrossRefBookmark::CrossRefBookmark(..)>" "- creation of cross-reference bookmark with an illegal PaM that does not expand over exactly one whole paragraph."' failed.

This is because in DocumentContentOperationsManager::MoveRange() the
flag bNullContent is set after the node has been split; in this case the
nContent is of course always 0.

Later the function then restores aSavePam to the index 0 of the next
node, when it actually shouldn't do anything because the JoinNext()
already positioned it correctly at the merge-index of the re-joined node.

(regression from f34af96b8043f43e90b7b36f4c5fef4a33b5902c)

Change-Id: I64d50e70b19e2fd81537a9771fa8706898b17642