tdf#98700 docxexport: don't force on ParaKeepTogether
The mere presence of SvxFormatKeepItem was ENABLING it during export,
without checking to see if it was actually turned on or off.
Both DOC and RTF check the value, and set accordingly, so do the
same for DOCX.

Merely toggling the setting on and off is enough to create the
property, so this is a nasty bug that only affects inquisitive
people.

Change-Id: I02d83a255f5b4ff8c5124302a52a3126dad40b67
Reviewed-on: https://gerrit.libreoffice.org/41309
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>