tdf#111524 sw: don't expand annotation mark when showing redline...
... at a position directly behind the annotation mark.

This happens with the the annotation "__Annotation__5847_848227920" in
the bugdoc, followed by <text:change text:change-id="ct266312216"/>.

When the redline is shown, the annotation mark expands from
5 - 101 to 5 - 240, but the SwPostItField is of course still at 101,
so during the ODF export lcl_FillAnnotationStartArray() asserts.

The problem is that the "bDelete" parameter of SwTextNode::Update()
disables a whole bunch of code that prevents update of positions
of redlines, bookmarks, fly frames, and shell cursors.

This was introduced in 2004 with CWS dvoea1, but there is absolutely
no documentation of what problem it was supposed to solve.
So just try to remove it here and see if it causes any issue.

Change-Id: I2d1f78c7163eddaf0ce6bbb7c6685ca759874ec5