drop unused SvxScriptTypeItem
apparently there was never a user for this item since initial commit
in the year 2000. SvxScriptSetItem gets the ScriptType by parameter
see GetItemOfScript and PutItemForScriptType.

remove SID_ATTR_CHAR_SCRIPTTYPE and don't put the item in any set.
keep the gap in the ID range because it's not part of any range
and there are already other gaps.

Change-Id: Iee24f13b329593af2dc5548d8b6bc2282c2dc5b3
Reviewed-on: https://gerrit.libreoffice.org/41423
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>