tdf#108765 DOCX: after import, allow default styles again
During import, we don't want LO's style defaults to intermingle
with those of MSO [1]. However, if any styles are NOT defined by
the document itself, then we want to use that new style with
appropriate defaults.

For example, if footnotes were never used, that style might not be
defined in the .docx file. Later, if a user inserts a footnote
for the first time, we certainly don't want the "default paragraph"
settings. Instead, we want the default "Footnote" settings.

[1]: see commit 3c4a2375c41398e01ff856d8ad96f8d70a918079

Change-Id: Ia71490e795885d3c16f764d0f716f6499765cb67
Reviewed-on: https://gerrit.libreoffice.org/41546
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>