Avoid deadlock under --enable-debug
...as seen during JunitTest_sc_unoapi_2 in an --enable-dbgutil build, where
thread 1 owns the SolarMutex and tries to acquire the scheduler mutex while
thread 7 owns the scheduler mutex and tries to re-acquire the SolarMutex.  Keep
the SAL_WARN output about unprocessed idles just in case somebody considers it
useful (but it needs to be taken with a grain of salt, of course).

Change-Id: Ib83d02876dc0d3b3b9a4c08581edf9c378fc911b