tdf#104937 RTF import: \trwWidthA is an absolute value
Not a relative one, so similar to \cellx need to convert these before
sending the value to dmapper (which expects an OOXML value, which is
always relative).

The result of the correct type is that the relative width of the last
cell in the table row is smaller, so other cells are wider -> they now
reach the expected width.

Change-Id: If282eb9a5ec644b38e552b5a66fc64a93538f216
Reviewed-on: https://gerrit.libreoffice.org/41668
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>