sw: WW8 import: properly handle consecutive range annotations
moz1187869-1.doc crashes on export to ODF because an annotation
mark has been inserted that does not have a SwPostItField.

In fact, the SwPostItField exists, but the annotation mark is at
the wrong position.

There are 2 consecutive range annotations in the file, so when
Read_And() calls MoveAttrs(), the end of the both annotations
are moved 2 times, hence later SwFltControlStack::SetAttrInDoc()
will insert 2 annotation marks with the same end position,
and the 2nd one will overwrite the name that the 1st one set
on the 2nd SwPostItField, while the 1st SwPostItField remains
with an empty name.

Ensure that the end of an annotation mark is no longer moved
once it's on the position of the field inserted in Read_And().

Crash is a regression from e8f7307b131977bb84bc582ec35b49f3f67e71b9.

Change-Id: I0fd661e22e51701de67227f9145a13c62b10143e