tdf#112109 sw: split section frames inside table cells, ignore nested tables
Commit 0640c594254d1060ec416706e54f02788b7f1aa4 (tdf#108524 sw: split
section frames inside table cells, non-split text frames, 2017-07-06)
added support for moving text frame masters to a next page inside
section-in-table frames. But the code in SwFrame::GetNextSctLeaf()
responsible for this is not up to nested tables, so don't try to split
sections in this case.

Otherwise we'll end up with frames which are marked as "in table", but
don't actually have have table frame parents, so we crash in
SwFrame::IsFootnoteAllowed() which assumes being in a table means a
table frame parent.

Change-Id: Iff19a4eda21a4dbfb9562dea7af8ec6767d47873
Reviewed-on: https://gerrit.libreoffice.org/41748
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>