MSVC 15.0: Fix configure to work with update 15.3
Update 15.3 renamed redist directory from 150 to 141 and introduced
intermediate build number directories, that are different from the
compiler build number directories.

While the compiler is located under:

  [...]/MSVC/14.11.25503/bin/HostX64/x64/cl.exe

the merge modules are located now under:

  VC/redist/MSVC/14.11.25325/MergeModules

Change the hard coded detection approach, to the dynamic one: list
all directories in suspected locations and check for existnace of
the artifacts we are looking for.

Change-Id: Ie040f290d1748ab4ec5c715b6636cc3733329e1c
Reviewed-on: https://gerrit.libreoffice.org/41922
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>
Tested-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>