android: fix some layout issues with new-document buttons
functionality to create new documents was added for tdf#106325 with
commit c62b5ee907fff8498b4ea04c4d97281add9b22c1
However the commit had some issues, most notably clipping of the shadow
effect and alignment issues to to scaling down (via the animation
effect, took me quite a while until I looked at that) with having a
centered pivot. I think it was due to those alignment issues that the
label was set to a fixed with - which is a problem for future
translation.
This commit:
* changes the animation to not scale down the initial views
* changes the FABs' setting to use compatPadding - this not only ensures
  enough room for the shadow, but also allows for consistent look across
  android versions. Side effect is that it needs more height, so
* the effect was adjusted accordingly, to shrink to/expand from the
  lower-right corner of the parent view, instead of "mid-air"
* the labels were given a proper background drawable with rounded
  corners and some padding for a (much more9 polished look
* formatting of the labels was split out to a style, for ease of
  tweaking the looks
* add tools:visibility statement to see the elements in layout editor
* fixed the strings for the labels (and ID) to some less awkward ones

Change-Id: Ifd2d90ec8d23e40603edff19c9af9e0b32090c9f
Reviewed-on: https://gerrit.libreoffice.org/42033
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>